CUT-4953: Primary Device User (#217)
* primaryUser param and validation

* Improve retry logic in Invoke-SystemPut

Refactored the retry mechanism to use a $retry flag instead of $success, ensuring retries only occur for specific network errors. Now throws the error message if the update fails after all retries, and logs more detailed error information.

* remove throw

* primarySystemUser - Invoke-SystemPut

Adds logic to set the primary system user in JumpCloud if the user was bound successfully and the PrimaryUser flag is specified. This ensures the correct user is designated as the primary system user during migration.

* primarySystemUser - SystemContext

* CUT-4956 vuln permission fix

* validation for PrimaryUser and AutoBindJCUser

Refactored parameter validation to throw a ValidationMetadataException when PrimaryUser is true and AutoBindJCUser is false. This enhances error handling and clarifies the dependency between these parameters.

* add cb_primaryUser

* primaryUser checkbox

* Enable primary user checkbox with autobind option

The primary user checkbox is now enabled and checked when the autobind JumpCloud user option is selected, and disabled/unchecked when it is not. Also updated window title to version 2.9.5.

* move cb_primaryuser

* fix primaryUser position

Moved the 'Primary User' checkbox up to align with other controls in the selection form UI for improved layout consistency.

* primaryUser tests

* docs

* primary user tests

* primaryUser with systemContextBinding

Enhanced validation logic for the 'PrimaryUser' parameter to require both 'AutoBindJCUser' and 'systemContextBinding' to be set to true, ensuring correct parameter combinations and preventing misconfiguration.

* combine conditions to ensure that they work

* primaryUser tests

* minor release

* include PrimaryUser

* include PrimaryUser

* Update ModuleChangelog.md

* fix logic for autobindjcuser and systemcontextbinding

* fix tests

* Fix JumpCloud connectivity validation condition

Corrects the conditional check from $reportStatus to $autobindJCUser when validating JumpCloud connectivity. Also ensures the systemDescription object is only built when $reportStatus is true.

* fix systemID

* update docs

* docs + Changelog

* update test to account for builds in mac/windows

* missing tests, form not passing in primaryUser selection, print input if form

* set primary user in tests

* release date

---------

Co-authored-by: Joe Workman <joe.workman@jumpcloud.com>